+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Tests for YAML input
 - [#25](https://github.com/devatherock/email-sender/issues/25): Custom error response for validation errors
 - [#14](https://github.com/devatherock/email-sender/issues/14): Tests for sends to bcc addresses
+- Tests for sends using embedded SMTP server
 
 ### Changed
 - Upgraded `spring-boot` to `3.0.12`